Added dependencies
Mod files just need to add a line for a dependent file
Example:
-- #dependency DomEnhanced1_80.dm

Anywhere in the file. This will have references be shared, ie, if you want to reference a dominions enhanced monster or spell ID and expect it to be a DE ID, then when the DE ID is changed, it'll be changed to match.